requirements.txt
amqp==2.4.2
anyjson==0.3.3
asgiref==1.1.2
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
channels==1.1.8
channels-api==0.4.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
daphne==1.4.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.6
django-celery-results==1.0.4
django-ckeditor==5.6.1
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.0
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-redis==4.10.0
django-rest-captcha==0.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
DjangoUeditor==1.8.143
drf-extensions==0.4.0
drf-haystack==1.8.4
et-xmlfile==1.0.1
future==0.17.1
html5lib==1.0.1
httplib2==0.12.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipaddress==1.0.22
itypes==1.1.0
jdcal==1.4
jieba==0.39
Jinja2==2.10
kombu==4.5.0
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
neobolt==1.7.4
neotime==1.7.4
odfpy==1.4.0
olefile==0.46
openpyxl==2.6.1
Pillow==4.3.0
prompt-toolkit==2.0.9
py2neo==4.2.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycryptodome==3.8.1
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.7.2
pyOpenSSL==19.0.0
pysolr==3.8.1
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
redis==3.2.1
requests==2.21.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9
tablib==0.13.0
tesseract==0.1.3
Twisted==19.2.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0








